requirements.txt
python-telegram-bot==21.0




